lights flickering and minor tweaks

+++ Code/of_v0.8.0_vs_release/apps/myApps/ReactiveSpace/src/GeoScene.cpp
@@ -67,7 +67,7 @@ void GeoScene::Render()
 	}
 }
 
-void GeoScene::Update(int timeScale)
+void GeoScene::Update(float timeScale)
 {
 	//update hands
 	GeoParticle* gp;
