+++ Code/of_v0.8.0_vs_release/apps/myApps/ReactiveSpace/src/GridScene.cpp
@@ -169,7 +169,7 @@ void GridScene::Render()
 	}
 }
 
-void GridScene::Update(int timeScale)
+void GridScene::Update(float timeScale)
 {
 	int i = 0;
 	BirdParticle* p;
